+++ Controllers/JVSpeechController.m
@@ -14,7 +14,7 @@
 		_speechQueue = [[NSMutableArray alloc] initWithCapacity:15];
 		_synthesizers = [[NSArray alloc] initWithObjects:[[[NSSpeechSynthesizer alloc] initWithVoice:nil] autorelease], [[[NSSpeechSynthesizer alloc] initWithVoice:nil] autorelease], [[[NSSpeechSynthesizer alloc] initWithVoice:nil] autorelease], nil];
 
-		for( unsigned i = 0; i < [_synthesizers count]; i++ )
+		for( NSUInteger i = 0; i < [_synthesizers count]; i++ )
 			[[_synthesizers objectAtIndex:i] setDelegate:self];
 	}
 
@@ -32,7 +32,7 @@
 }
 
 - (void) startSpeakingString:(NSString *) string usingVoice:(NSString *) voice {
-	for( unsigned i = 0; i < [_synthesizers count]; i++ ) {
+	for( NSUInteger i = 0; i < [_synthesizers count]; i++ ) {
 		NSSpeechSynthesizer *synth = [_synthesizers objectAtIndex:i];
 		if( ! [synth isSpeaking] ) {
 			[synth setVoice:voice];
